+++ index.js
@@ -365,12 +365,7 @@ module.exports = {
 
         return null;
       } catch (err) {
-        self.apos.utils.error(err);
-
-        // We don't want to throw an error simply because emails didn't send.
-        self.apos.notify(req, 'There was an error sending email notifications.', {
-          type: 'error'
-        });
+        self.apos.utils.error('⚠️ apostrophe-forms submission email notification error: ', err);
 
         return null;
       }